Refactor TypeTestInsn class
